backend/requirements.txt
pyswisseph>=2.10.0
fastapi>=0.95.1
uvicorn>=0.22.0
numpy>=1.24.3
pandas>=2.0.1
requests>=2.31.0
yfinance>=0.2.18
python-dotenv>=1.0.0
pydantic>=2.0.0
firebase-admin>=6.2.0
pytest>=7.3.1
httpx>=0.24.1 